# backend/dist/swagger.yaml
basePath: /api/v1
consumes:
    - application/json
...
info:
    title: libroteca-angy
    version: 0.1.0
    description: 'Initial server using node.js SDK APIs'
    license:
        name: Apache-2.0
paths:
    '/items/searchUser/{user}':
        get:
            operationId: SearchUser
            produces:
                - application/json
            responses:
                '200':
                    description: ""
                    schema:
                        type: object
            tags:
                - Items
            security: []
            parameters:
                -
                    in: path
                    name: user
                    required: true
                    type: string
    '/items/searchEmail/{email}':
        get:
            operationId: SearchEmail
            produces:
                - application/json
            responses:
                '200':
                    description: ""
                    schema:
                        type: object
            tags:
                - Items
            security: []
            parameters:
                -
                    in: path
                    name: email
                    required: true
                    type: string
...
    '/items/borrarLeido/{idLibro}':
        get:
            operationId: DeleteReadedBook
            produces:
                - application/json
            responses:
                '200':
                    description: ""
                    schema:
                        type: object
            tags:
                - Items
            security: []
            parameters:
                -
                    in: path
                    name: idLibro
                    required: true
                    format: double
                    type: number
...
    '/items/addFavorito/{id}':
        get:
            operationId: AddFavorito
            produces:
                - application/json
            responses:
                '200':
                    description: ""
                    schema:
                        type: object
            tags:
                - Items
            security: []
            parameters:
                -
                    in: path
                    name: id
                    required: true
                    format: double
                    type: number
    '/items/borrarFavorito/{idLeido}':
        get:
            operationId: BorrarFavorito
            produces:
                - application/json
            responses:
                '200':
                    description: ""
                    schema:
                        type: object
            tags:
                - Items
            security: []
            parameters:
                -
                    in: path
                    name: idLeido
                    required: true
                    format: double
                    type: number
...
produces:
    - application/json
swagger: '2.0'
securityDefinitions:
    jwt:
        type: apiKey
        name: x-access-token
        in: header
host: 'localhost:3000'